[Wizard- Far Install/Create] Change Manifest file name to "FrameworkArchiveManifest.xml". Modify Md5 format in FrameworkArchiveMainfest.xml file. Add Md5 verification when install FAR. Fix the bug of Far installation can't work with Sun JDK 1.5.0_04 version.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@1029 6f19259b-4bc3-4df7-8a09-765794883524

+++ Tools/Source/FrameworkWizard/src/org/tianocore/frameworkwizard/far/Far.java
@@ -226,6 +226,12 @@ public class Far {
                         byte[] totalBuffer = new byte[(int) tempFile.length()];
                         FileInputStream fis = new FileInputStream(tempFile);
                         fis.read(totalBuffer);
+                        //
+                        //  Check Md5
+                        //
+                        if (!ffItem.getMd5Value().equalsIgnoreCase(FarMd5.md5(totalBuffer))){
+                            throw new Exception (ffItem.getRelativeFilename() + " Md5 is invaild!");
+                        }
                         outputStream.write(totalBuffer);
                         fis.close();
                         tempFile.delete();
